requirements.txt
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
torch>=2.0.0
torchvision>=0.15.0
flask==3.0.0
pyyaml==6.0.1
matplotlib==3.8.2

pandas==2.3.3
pytz==2025.2
tzdata==2025.3
